Added natHacks Brainflow Notebook
I attempted to connect to the muse via this code but unfortunately it doesn't like windows PC's. If someone could try using bluetooth on a mac (after changing the port properly) That would be great.

The readme has a great video on how to use brainflow, if it doesn't work out though we can just pay $15 for mind monitor.